Split DBRepository in two separate ones, for now
