+++ logs/app.log
@@ -1,47 +1,47 @@
-2025-07-01 12:41:45.151 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
-2025-07-01 12:41:45.151 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 12:41:45.151 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-01 12:41:45.151 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-01 12:41:45.174 INFO  [pool-1-thread-1] org.example.Main - User johnny logged in
 2025-07-01 14:58:19.696 ERROR [pool-1-thread-1] o.example.Services.CustomerServices - the money weren't received, it returned false
 2025-07-01 14:58:19.698 INFO  [pool-1-thread-1] org.example.Main - User johnny sent the money
 2025-07-01 14:58:19.733 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were sent
 2025-07-01 14:58:19.733 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were received
-2025-07-01 14:58:19.762 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 14:58:19.762 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-01 14:58:19.775 INFO  [pool-1-thread-2] org.example.Main - User johnny sent the money
 2025-07-01 15:02:53.411 ERROR [pool-1-thread-1] o.example.Services.CustomerServices - the money weren't received, it returned false
 2025-07-01 15:02:53.449 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were sent
 2025-07-01 15:02:53.451 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were received
-2025-07-01 15:02:53.467 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 15:02:53.467 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-01 15:04:32.451 ERROR [pool-1-thread-2] o.example.Services.CustomerServices - the money weren't received, it returned false
 2025-07-01 15:04:32.475 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money were sent
 2025-07-01 15:04:32.477 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money were received
-2025-07-01 15:04:32.507 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 15:04:32.507 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-01 15:26:37.111 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money were sent
 2025-07-01 15:26:37.110 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were sent
 2025-07-01 15:26:37.115 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money were received
 2025-07-01 15:26:37.115 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money were received
-2025-07-01 15:26:37.144 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
-2025-07-01 15:26:37.144 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 15:26:37.144 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-01 15:26:37.144 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-01 15:29:53.132 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was sent
 2025-07-01 15:29:53.132 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was sent
 2025-07-01 15:29:53.137 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was received
 2025-07-01 15:29:53.137 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was received
-2025-07-01 15:29:53.156 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
-2025-07-01 15:29:53.156 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
+2025-07-01 15:29:53.156 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-01 15:29:53.156 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-03 10:16:44.599 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was sent
 2025-07-03 10:16:44.599 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was sent
 2025-07-03 10:16:44.606 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was received
 2025-07-03 10:16:44.606 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was received
-2025-07-03 10:16:44.636 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
-2025-07-03 10:16:44.636 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
+2025-07-03 10:16:44.636 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-03 10:16:44.636 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-03 13:14:43.198 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was sent
 2025-07-03 13:14:43.198 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was sent
 2025-07-03 13:14:43.201 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was received
 2025-07-03 13:14:43.203 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was received
-2025-07-03 13:14:43.222 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
-2025-07-03 13:14:43.222 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
+2025-07-03 13:14:43.222 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-03 13:14:43.222 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
 2025-07-03 19:51:36.791 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was sent
 2025-07-03 19:51:36.791 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was sent
 2025-07-03 19:51:36.796 INFO  [pool-1-thread-2] o.example.Services.CustomerServices - money was received
 2025-07-03 19:51:36.796 INFO  [pool-1-thread-1] o.example.Services.CustomerServices - money was received
-2025-07-03 19:51:36.836 INFO  [pool-1-thread-2] org.example.DBRepository - Number of rows inserted: 1
-2025-07-03 19:51:36.836 INFO  [pool-1-thread-1] org.example.DBRepository - Number of rows inserted: 1
+2025-07-03 19:51:36.836 INFO  [pool-1-thread-2] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
+2025-07-03 19:51:36.836 INFO  [pool-1-thread-1] org.example.Repositories.TransfersRepository - Number of rows inserted: 1
